+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * Content handler callbacks (like `to_html`) can return `{:file,
     filename}` for the body, which will use `Plug.Conn.send_file/3` to
     send the response
+  * The resource macro can restrict route matches to a specific host.
 
 * Backwards incompatible changes
   * The default values of `expires`, `generate_etag`, and